+++ test/instructions/from.test.ts
@@ -30,6 +30,17 @@ describe("FROM", () => {
         assert.equal(from.getBuildStage(), null);
     });
 
+    it("FROM no${de:-de}", () => {
+        let dockerfile = DockerfileParser.parse("FROM no${de:-de}");
+        let from = dockerfile.getFROMs()[0];
+        assert.equal(from.getImage(), "no${de:-de}");
+        assert.equal(from.getImageName(), "no${de:-de}");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 16);
+        assert.equal(from.getImageTagRange(), null);
+        assert.equal(from.getImageDigestRange(), null);
+        assert.equal(from.getBuildStage(), null);
+    });
+
     it("FROM node:", () => {
         let dockerfile = DockerfileParser.parse("FROM node:");
         let from = dockerfile.getFROMs()[0];
@@ -48,6 +59,17 @@ describe("FROM", () => {
         assert.equal(from.getImageName(), "node");
         assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assertRange(from.getImageTagRange(), 0, 10, 0, 16);
+        assert.equal(from.getImageDigestRange(), null);
+        assert.equal(from.getBuildStage(), null);
+    });
+
+    it("FROM node:${tag:-latest}", () => {
+        let dockerfile = DockerfileParser.parse("FROM node:${tag:-latest}");
+        let from = dockerfile.getFROMs()[0];
+        assert.equal(from.getImage(), "node:${tag:-latest}");
+        assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
+        assertRange(from.getImageTagRange(), 0, 10, 0, 24);
         assert.equal(from.getImageDigestRange(), null);
         assert.equal(from.getBuildStage(), null);
     });
@@ -71,6 +93,17 @@ describe("FROM", () => {
         assertRange(from.getImageNameRange(), 0, 5, 0, 9);
         assert.equal(from.getImageTagRange(), null);
         assertRange(from.getImageDigestRange(), 0, 10, 0, 81);
+        assert.equal(from.getBuildStage(), null);
+    });
+
+    it("FROM node@${digest:-sha256:613685c22f65d01f2264bdd49b8a336488e14faf29f3ff9b6bf76a4da23c4700}", () => {
+        let dockerfile = DockerfileParser.parse("FROM node@${digest:-sha256:613685c22f65d01f2264bdd49b8a336488e14faf29f3ff9b6bf76a4da23c4700");
+        let from = dockerfile.getFROMs()[0];
+        assert.equal(from.getImage(), "node@${digest:-sha256:613685c22f65d01f2264bdd49b8a336488e14faf29f3ff9b6bf76a4da23c4700");
+        assert.equal(from.getImageName(), "node");
+        assertRange(from.getImageNameRange(), 0, 5, 0, 9);
+        assert.equal(from.getImageTagRange(), null);
+        assertRange(from.getImageDigestRange(), 0, 10, 0, 91);
         assert.equal(from.getBuildStage(), null);
     });